Update running instances
Whenever ebin-bubbl is run, if there already is a running instance, it
will receive the new command and the old instance will exit. This solves
the problem of either having to kill running ebin-bubbles before
starting a new one, or letting many instances run on top of each other.

The command is sent through a FIFO file, the path to which is
configurable in the config.h. The file is removed when the program
exits. If the program starts and the file exists, it is assumed that
there is a running program that can receive new commands.